Add declared crate features to the Rust sync tool (#573)
## Usage and product changes
Introduce `declared_features` to the Rust sync tool, parsed from the
tags of Bazel `rust_` targets, to generate `Cargo.toml` files with
predeclared features that are not automatically used in the dependent
targets (this happens with `crate_features`).

## Motivation
We need a way to declare features without their automatic usage to be
able to activate them manually through `cargo` only when needed.

## Implementation
Parse `tags` in order to search for a new tag containing a list of
declared features. Save these features into a new separate field which
is then passed to the crate features declaration.