Add base64 crate dependency (#576)
## Usage and product changes
Add the latest version of the `base64` crate to the Rust dependencies.
Additionally, pin the `backtrace` to the previous version to support old
versions of Rust (its newest version, which is auto-updated in cargo,
[requires](https://crates.io/crates/backtrace/versions) Rust 1.82, and
it breaks our applications builds).

## Motivation
This crate enables us to encode byte arrays into strings with a total
space overhead of ~1.33x, making it useful for transferring TypeDB IIDs
over the network when human readability is not required.

## Implementation